+++ tencentcloud/thpc/v20220401/errors.go
@@ -17,11 +17,23 @@ package v20220401
 const (
 	// 此产品的特有错误码
 
+	// CAM签名/鉴权错误。
+	AUTHFAILURE = "AuthFailure"
+
 	// DryRun 操作，代表请求将会是成功的，只是多传了 DryRun 参数。
 	DRYRUNOPERATION = "DryRunOperation"
 
+	// 操作失败。
+	FAILEDOPERATION = "FailedOperation"
+
+	// 内部错误。
+	INTERNALERROR = "InternalError"
+
 	// cvm调用失败。
 	INTERNALERROR_CALLCVM = "InternalError.CallCvm"
+
+	// 参数错误。
+	INVALIDPARAMETER = "InvalidParameter"
 
 	// 参数格式有误。
 	INVALIDPARAMETER_MALFORMED = "InvalidParameter.Malformed"
@@ -50,11 +62,29 @@ const (
 	// 参数值过小。
 	INVALIDPARAMETERVALUE_TOOSMALL = "InvalidParameterValue.TooSmall"
 
+	// 参数值重复。不支持此操作。
+	INVALIDPARAMETERVALUE_VALUEDUPLICATED = "InvalidParameterValue.ValueDuplicated"
+
+	// 超过配额限制。
+	LIMITEXCEEDED = "LimitExceeded"
+
 	// 缺少参数错误。
 	MISSINGPARAMETER = "MissingParameter"
 
+	// 操作被拒绝。
+	OPERATIONDENIED = "OperationDenied"
+
+	// 请求的次数超过了频率限制。
+	REQUESTLIMITEXCEEDED = "RequestLimitExceeded"
+
+	// 资源被占用。
+	RESOURCEINUSE = "ResourceInUse"
+
 	// 资源不足。
 	RESOURCEINSUFFICIENT = "ResourceInsufficient"
+
+	// 资源不存在。
+	RESOURCENOTFOUND = "ResourceNotFound"
 
 	// 无法找到弹性伸缩组ID。
 	RESOURCENOTFOUND_AUTOSCALINGGROUPID = "ResourceNotFound.AutoScalingGroupId"
@@ -71,6 +101,15 @@ const (
 	// 无法找到ID对应节点。
 	RESOURCENOTFOUND_NODEID = "ResourceNotFound.NodeId"
 
+	// 未授权操作。
+	UNAUTHORIZEDOPERATION = "UnauthorizedOperation"
+
+	// 未知参数错误。
+	UNKNOWNPARAMETER = "UnknownParameter"
+
+	// 操作不支持。
+	UNSUPPORTEDOPERATION = "UnsupportedOperation"
+
 	// 该伸缩组已绑定集群，请更换伸缩组。
 	UNSUPPORTEDOPERATION_AUTOSCALINGGROUPALREADYBINDED = "UnsupportedOperation.AutoScalingGroupAlreadyBinded"
 
@@ -85,6 +124,12 @@ const (
 
 	// 节点状态不支持此操作。
 	UNSUPPORTEDOPERATION_NODESTATUSNOTSUPPORT = "UnsupportedOperation.NodeStatusNotSupport"
+
+	// 参数值过大，不支持此操作。
+	UNSUPPORTEDOPERATION_PARAMETERTOOLARGE = "UnsupportedOperation.ParameterTooLarge"
+
+	// 参数值过小，不支持此操作。
+	UNSUPPORTEDOPERATION_PARAMETERTOOSMALL = "UnsupportedOperation.ParameterTooSmall"
 
 	// vpc冲突，不支持当前操作。
 	UNSUPPORTEDOPERATION_VPCIDCONFLICT = "UnsupportedOperation.VpcIdConflict"